requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# AI/ML
transformers>=4.35.0
torch>=2.1.0
anthropic>=0.7.0

# Data Sources
newsapi-python>=0.2.7
tweepy>=4.14.0
alpha-vantage>=2.3.1

# Blockchain (Story Protocol)
web3>=6.11.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
